requirements.txt
asn1crypto==1.5.1
boto3==1.37.29
botocore==1.37.29
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
filelock==3.18.0
idna==3.10
jmespath==1.0.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
platformdirs==4.3.7
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
tomlkit==0.13.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
